Move events & inventory types to metric pkg
